+++ .github/workflows/test_indicators.yml
@@ -46,4 +46,4 @@ jobs:
       - name: Pytest
         run: |
           . $HOME/.poetry/env
-          poetry run pytest tests/general_indicator_tests
+          poetry run pytest tests/general_indicator_tests/index_type_tests.py
